Add a "Branches" submenu for switching branches
This removes the "Switch branches" and "Create branch" menu items,
instead giving each branch a dedicated menu item in a "Branches"
submenu and placing "Create branch" at the end of it.

If loading of branches is too slow for frequent use in a menu
(longer than a second), then the branch list and currently
checked-out branch are cached in the repository folder's attributes
and refreshed after 10 minutes as necessary.